chore(engine): address code review points
- Rollback changes to DeploymentTest
- Move shouldSkipCustomListenersOnProcessInstanceModification to ProcessInstanceModificationTest
- Adjust then comment
- Remove unused variable
